Add handlers inside handlers folder
