+++ src/actions/handlers/openModalHandler.ts
@@ -1,4 +1,4 @@
-import { IOpenModal } from "../types/types";
+import { IOpenModal } from "../../types/types";
 
 function openModalHandler({ modalName, openModal }: IOpenModal) {
   openModal(modalName);
