+++ src/actions/handlers/registerHandler.ts
@@ -1,6 +1,6 @@
 import { Keypair } from "stellar-sdk";
-import setKeypair from "./setKeypair";
-import { ISetKeyPair } from "../types/types";
+import setKeypair from "../stellar/setKeypair";
+import { ISetKeyPair } from "../../types/types";
 
 function registerHandler({
   setSecretKey,
