+++ src/components/modals/home/StepOneLogin.tsx
@@ -2,8 +2,8 @@ import React, { useState } from "react";
 import Button from "../../commons/Button";
 import { ILoginModalStepOne } from "../../../types/types";
 import { Checkbox } from "antd";
-import closeModalHandler from "../../../actions/closeModalHandler";
-import openModalHandler from "../../../actions/openModalHandler";
+import closeModalHandler from "../../../actions/handlers/closeModalHandler";
+import openModalHandler from "../../../actions/handlers/openModalHandler";
 import { Modal } from "antd";
 import { handleCheck } from "../../../actions/utils";
